refactor: move AboutSection.tsx to components folder
